Crate cardano_chain_follower

Cardano chain follower.

Modules

follow 🔒

Cardano chain follow module.

mithril_snapshot 🔒

Internal Mithril snapshot functions.

Structs

Follower

Cardano chain follower.

FollowerConfig

Configuration for the Cardano chain follower.

FollowerConfigBuilder

Builder used to create FollowerConfigs.

MultiEraBlockData

CBOR encoded data of a multi-era block.

ReadBlock

Handler for receiving the read block response from the client.

ReadBlockRange

Handler for receiving the read block range response from the client.

Enums

ChainUpdate

Enum of chain updates received by the follower.

Error

Crate error type.

Network

Enum of possible Cardano networks.

Point

A point within a chain

PointOrTip

A point in the chain or the tip.

Constants

MAINNET_NAME 🔒

The human readable name of the Cardano mainnet network.

PREPROD_NAME 🔒

The human readable name of the Cardano pre-production network.

PREVIEW_NAME 🔒

The human readable name of the Cardano preview network.

TESTNET_NAME 🔒

The human readable name of a Cardano local testnet network.

Functions

network_genesis_values

Return genesis values for given network

validate_multiera_block 🔒

Validate a multi-era block.

Type Aliases

Result

Crate result type.