Module cardano_chain_follower::chain_sync

Sync from the chain to an in-memory buffer.

All iteration of the chain is done through this buffer or a mithril snapshot. Consumers of this library do not talk to the node directly.

Constants

MAX_NODE_CONNECT_RETRIES 🔒

The maximum number of times we wait for a nodeChainUpdate to connect. Currently set to never give up.

MAX_NODE_CONNECT_TIME_SECS 🔒

The maximum number of seconds we wait for a node to connect.

PEER_FAILURE_RECONNECT_DELAY 🔒

How long we wait before trying to reconnect to a peer when it totally fails our attempts.

Functions

chain_sync 🔒

Handle the background downloading of Mithril snapshots for a given network. Note: There can ONLY be at most three of these running at any one time. This is because there can ONLY be one snapshot for each of the three known Cardano networks.

fetch_block_from_peer 🔒

Fetch a single block from the Peer, and Decode it.

follow_chain 🔒

Follows the chain until there is an error. If this returns it can be assumed the client is disconnected.

live_sync_backfill 🔒

Backfill the live chain, based on the Mithril Sync updates. This does NOT return until the live chain has been backfilled from the end of mithril to the current synced tip blocks.

live_sync_backfill_and_purge 🔒

Backfill and Purge the live chain, based on the Mithril Sync updates.

persistent_reconnect 🔒

Do not return until we have a connection to the peer.

point_at_tip 🔒

Is the current point aligned with what we know as tip.

process_next_block 🔒

Process a rollback detected from the peer.

process_rollback 🔒

Process a rollback detected from the peer.

process_rollback_actual 🔒

Process a rollback.

purge_and_intersect_tip 🔒

Purge the live chain, and intersect with TIP.

resync_live_tip 🔒

Resynchronize to the live tip in memory.

retry_connect 🔒

Try and connect to a node, in a robust and quick way.