Module cardano_chain_follower::chain_sync_live_chains

Storage of each Live Chain per Blockchain.

Structs

ProtectedLiveChainBlockList 🔒

Because we have multi-entry relationships in the live-chain protect it with a read/write lock. The underlying SkipMap is still capable of multiple simultaneous reads from multiple threads which is the most common access.

Constants

DATA_RACE_BACKOFF_SECS 🔒

Number of seconds to wait if we detect a SyncReady race condition.

Statics

LIVE_CHAINS 🔒

Handle to the mithril sync thread. One for each Network ONLY.

PEER_TIP 🔒

Latest TIP received from the Peer Node.

Functions

find_best_fork_block 🔒

Find best block from a fork relative to a point.

get_fill_to_point 🔒

Get the fill tp point for a chain.

get_intersect_points 🔒

Get intersection points to try and find best point to connect to the node on reconnect.

get_live_block 🔒

Get the Live block relative to the specified point. The starting block must exist if the search is strict.

get_live_chain 🔒

Get the LiveChainBlockList for a particular Network.

get_live_head_point 🔒

Get the head Point currently in the live chain.

get_peer_tip 🔒

Set the last TIP received from the peer.

live_chain_add_block_to_tip 🔒

Insert a block into the live chain (in-order). Can ONLY be used to add a new tip block to the live chain. rollback_count should be set to 1 on the very first connection, after that, it is maintained by this function, and MUST not be modified elsewhere.

live_chain_backfill 🔒

Backfill the live chain with the block set provided.

live_chain_length 🔒

Get the length of the live chain. Probably used by debug code only, so its ok if this is not use.

purge_live_chain 🔒

On an immutable update, purge the live-chain up to the new immutable tip. Will error if the point is not in the Live chain.

update_peer_tip 🔒

Set the last TIP received from the peer.

Type Aliases

LiveChainBlockList 🔒

Type we use to manage the Sync Task handle map.