Module cardano_chain_follower::chain_sync_ready

Flag to control if chain sync for a blockchain is ready. Can not consume the blockchain data until it is.

Structs

SyncReady 🔒

Data we hold related to sync being ready or not.

SyncReadyWaiter 🔒

Waiter for sync to become ready, use signal when it is.

Constants

SYNC_READY_RACE_BACKOFF_SECS 🔒

Number of seconds to wait if we detect a SyncReady race condition.

Statics

SYNC_READY 🔒

Lock to prevent using any blockchain data for a network UNTIL it is synced to TIP. Pre-initialized for all possible blockchains, so it’s safe to use expect to access a value.

Functions

block_until_sync_ready 🔒

Block until the chain is synced to TIP. This is necessary to ensure the Blockchain data is fully intact before attempting to consume it.

check_sync_ready 🔒

Get a Read lock on the Sync State, and return if we are ready or not.

get_chain_update_rx_queue 🔒

Get the Broadcast Receive queue for the given chain updates.

get_chain_update_tx_queue 🔒

Get the Broadcast Transmit queue for the given chain updates.

notify_follower 🔒

Sand a chain update to any subscribers that are listening.

wait_for_sync_ready 🔒

Write Lock the SYNC_READY lock for a network. When we are signaled to be ready, set it to true and release the lock.