cardano_chain_follower

Module follow

Cardano chain follow module.

Modules

task 🔒

Contains functions related to the Follower’s background task.

Structs

ClientConnectInfo 🔒

Information used to connect to a client.

Follower

Cardano chain follower.

FollowerConfig

Configuration for the Cardano chain follower.

FollowerConfigBuilder

Builder used to create FollowerConfigs.

ReadBlock

Handler for receiving the read block response from the client.

ReadBlockRange

Handler for receiving the read block range response from the client.

Enums

ChainUpdate

Enum of chain updates received by the follower.

Constants

DEFAULT_CHAIN_UPDATE_BUFFER_SIZE 🔒

Default Follower block buffer size.

Functions

read_block_from_network 🔒

Reads a block from the network using the N2N client.

read_block_range_from_network 🔒

Reads a range of blocks from the network using the N2N client.

resolve_tip 🔒

Finds the tip point.

set_client_read_pointer 🔒

Sets the N2N remote client’s read pointer.