Module cardano_chain_follower::follow::follow_task

Contains functions related to the Follower’s background task.

Structs

MithrilSnapshotState 🔒

Holds the state of Mithril snapshot functions in the follow task.

TaskState 🔒

Holds the locks and channels used by the follow task.

Enums

Request

Follow task’s requests.

Response

Follow task’s responses.

Functions

handle_request 🔒

Handles a request.

next_from_client 🔒

Waits for the next update from the node the client is connected to.

run 🔒

Runs a Follower background task.

send_next_chain_update 🔒

Sends the next chain update to the follower. This can be either read from the Mithril snapshot (if configured) or from the N2N remote client.

set_client_read_pointer 🔒

Sets the N2N remote client’s read pointer.