Module cardano_chain_follower::metadata

Metadata decoding and validating.

Modules

cip36

Decoder and Validator for CIP36 Metadata

cip509

Cardano Improvement Proposal 509 (CIP-509) metadata module. Doc Reference: https://github.com/input-output-hk/catalyst-CIPs/tree/x509-envelope-metadata/CIP-XXXX CDDL Reference: https://github.com/input-output-hk/catalyst-CIPs/blob/x509-envelope-metadata/CIP-XXXX/x509-envelope.cddl

raw_aux_data 🔒

Raw Auxiliary Data Decoding

Structs

DecodedMetadata 🔒

Decoded Metadata for a single transaction. The key is the Primary Label of the Metadata.
For example, CIP15/36 uses labels 61284 & 61285, 61284 is the primary label, so decoded metadata will be under that label.

DecodedMetadataItem

An individual decoded metadata item.

DecodedTransaction

Decoded Metadata for a all transactions in a block. The Key for both entries is the Transaction offset in the block.

Enums

DecodedMetadataValues

Possible Decoded Metadata Values. Must match the key they relate too, but the consumer needs to check this.

Type Aliases

ValidationReport

List of all validation errors (as strings) Metadata is considered Valid if this list is empty.