Module cardano_chain_follower::mithril_snapshot_config

Configuration for the Mithril Snapshot used by the follower.

Structs

MithrilSnapshotConfig

Configuration used for the Mithril Snapshot downloader.

MithrilUpdateMessage 🔒

Message we send when Mithril Snapshot updates

Constants

TMP_SUB_DIR 🔒

Subdirectory where we unpack archives temporarily.

Statics

SYNC_JOIN_HANDLE_MAP 🔒

Handle to the mithril sync thread. One for each Network ONLY.

Functions

check_writable 🔒

Check that a given mithril snapshot path and everything in it is writable. We don’t care why its NOT writable, just that it is either all writable, or not. Will return false on the first detection of a read only file or directory.

is_hex 🔒

Check if a string is an even number of hex digits.

remove_whitespace 🔒

Remove whitespace from a string and return the new string

Type Aliases

SyncMap 🔒

Type we use to manage the Sync Task handle map.