Module cardano_chain_follower::mithril_snapshot_sync

Internal Mithril snapshot downloader task.

This task is responsible for downloading Mithril snapshot files. It downloads the latest snapshot file and then sleeps until the next snapshot is available.

Macros

next_iteration 🔒

Cleanup the client explicitly and do a new iteration of the loop.

Enums

SnapshotStatus 🔒

Status of checking if we have a new snapshot to get or not.

Constants

DOWNLOAD_ERROR_RETRY_DURATION 🔒

We shouldn’t get errors that need to wait for this, but if we do wait this long. These errors should be transient if they occur.

EXPECTED_MITHRIL_UPDATE_CHECK_INTERVAL 🔒

Average Mithril Update is 6 Hrs, so don’t wait longer than 7.

MAXIMUM_MITHRIL_UPDATE_CHECK_INTERVAL 🔒

Average Mithril Update is 6 Hrs, so don’t wait longer than 7.

MINIMUM_MITHRIL_UPDATE_CHECK_DURATION 🔒

The minimum duration between checks for a new Mithril Snapshot. (Must be same as MINIMUM_MITHRIL_UPDATE_CHECK_INTERVAL)

MINIMUM_MITHRIL_UPDATE_CHECK_INTERVAL 🔒

The minimum duration between checks for a new Mithril Snapshot. (Must be same as MINIMUM_MITHRIL_UPDATE_CHECK_DURATION)

MITHRIL_IMMUTABLE_SUB_DIRECTORY 🔒

Relative Directory for Immutable data within a full mithril snapshot.

Functions

background_mithril_update 🔒

Handle the background downloading of Mithril snapshots for a given network. Note: There can ONLY be at most three of these running at any one time. This is because there can ONLY be one snapshot for each of the three known Cardano networks.

background_validate_mithril_snapshot 🔒

Start Mithril Validation in the background, and return a handle so we can check when it finishes.

calculate_sleep_duration 🔒

Calculate how long we should wait before we check for another Mithril snapshot.

check_snapshot_to_download 🔒

Check if we have a new snapshot to download, and if so, return its details.

chunk_filename_to_chunk_number 🔒

Convert a chunk filename into its numeric equivalent.

cleanup 🔒

We can accumulate junk depending on errors or when we terminate, make sure we are always clean.

connect_client 🔒

This function creates a client based on the given network and genesis vkey.

create_client 🔒

Create a client, should never fail, but return None if it does, because we can’t continue.

download_and_validate_snapshot 🔒

Downloads and validates a snapshot from the aggregator.

download_and_verify_snapshot_certificate 🔒

Download and Verify the Snapshots certificate

get_latest_snapshots 🔒

Returns the Latest and chronologically previous snapshots data from the Aggregator. Will return None if it can not get the Snapshot list, or there are no entries in it. If there is only a single entry then the latest and chronologically next will be identical.

get_latest_validated_mithril_snapshot 🔒

See if we have a latest snapshot already, and if so, validate it.

get_mithril_snapshot_and_certificate 🔒

Get the Snapshot Data itself from the Aggregator, and a validate Certificate.

get_mithril_tip 🔒

Get the tip block from the Immutable chain, and the block immediately proceeding it.

get_snapshot 🔒

Get the actual snapshot from the specified snapshot_item from the list of snapshots. Returns None if there are any issues doing this, otherwise the Snapshot. The only issues should be transient communications errors.

get_snapshot_by_id 🔒

Given a particular snapshot ID, find the Actual Snapshot for it.

recover_existing_snapshot 🔒

Get the Mithril client and recover our existing mithril snapshot data, if any.

sleep_until_next_probable_update 🔒

Sleep until its likely there has been another mithril snapshot update.

trim_chunk_list 🔒

Remove any chunks from the chunk list which exceed the max_chunk.

validate_mithril_snapshot 🔒

Validate that a Mithril Snapshot downloaded matches its certificate.