Module cardano_chain_follower::mithril_turbo_downloader

Turbo Downloads for Mithril Snapshots.

Macros

changed_file 🔒

This macro is what happens every time the file is different from previous.

new_file 🔒

This macro is what happens every time we decide the file can’t be deduplicated.

Structs

Inner

A snapshot downloader that accelerates Download using aria2.

MithrilTurboDownloader

A snapshot downloader that accelerates Download using aria2.

Functions

get_file_size_sync 🔒

Get the size of a particular file. None = failed to get size (doesn’t matter why).

mmap_open_sync 🔒

Open a file using mmap for performance.