Module cardano_chain_follower::multi_era_block_data::ouroboros_impl_self_referenced_multi_era_block

Encapsulates implementation details for a self-referencing struct. This module is only visible when using –document-private-items.

Structs

BorrowedFields 🔒

A struct for holding immutable references to all tail and immutably borrowed fields in an instance of SelfReferencedMultiEraBlock.

BorrowedMutFields 🔒

A struct for holding mutable references to all tail fields in an instance of SelfReferencedMultiEraBlock.

Heads 🔒

A struct which contains only the head fields of SelfReferencedMultiEraBlock.

SelfReferencedMultiEraBlock 🔒

The self-referencing struct.

SelfReferencedMultiEraBlockAsyncBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of making your code both easier to refactor and more readable. Call build() to construct the actual struct. The fields of this struct should be used as follows:

SelfReferencedMultiEraBlockAsyncSendBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of making your code both easier to refactor and more readable. Call build() to construct the actual struct. The fields of this struct should be used as follows:

SelfReferencedMultiEraBlockAsyncSendTryBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of making your code both easier to refactor and more readable. Call try_build() or try_build_or_recover() to construct the actual struct. The fields of this struct should be used as follows:

SelfReferencedMultiEraBlockAsyncTryBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of making your code both easier to refactor and more readable. Call try_build() or try_build_or_recover() to construct the actual struct. The fields of this struct should be used as follows:

SelfReferencedMultiEraBlockBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of making your code both easier to refactor and more readable. Call build() to construct the actual struct. The fields of this struct should be used as follows:

SelfReferencedMultiEraBlockInternal 🔒

SelfReferencedMultiEraBlockTryBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of making your code both easier to refactor and more readable. Call try_build() or try_build_or_recover() to construct the actual struct. The fields of this struct should be used as follows:

Functions

check_if_okay_according_to_checkers 🔒

type_asserts 🔒