Module cardano_chain_follower::stats

Cardano Chain Follower Statistics

Structs

Follower

Individual Follower stats

Live

Statistics related to the live blockchain

Mithril

Statistics related to Mithril Snapshots

Rollback

Statistics related to a single depth of rollback

RollbackTypeIter

An iterator over the variants of RollbackType

Rollbacks

Statistics for all our known rollback types Rollback Vec is sorted by depth, ascending.

Statistics

Statistics for a single follower network.

Enums

MithrilSyncFailures 🔒

Mithril Sync Failures.

MithrilValidationState 🔒

State of the Mithril cert validation.

RollbackType

The types of rollbacks we track for a chain.

Statics

ROLLBACKS_MAP 🔒

Statistics of rollbacks detected per chain.

STATS_MAP 🔒

The statistics being maintained per chain.

Functions

backfill_ended 🔒

When did we start the backfill.

backfill_started 🔒

When did we start the backfill.

lookup_rollback_map 🔒

Get the actual rollback map for a chain.

lookup_stats 🔒

Get the stats for a particular chain.

mithril_dl_finished 🔒

Record when DL finished, if it fails, set size to None, otherwise the size of the downloaded file.

mithril_dl_started 🔒

Record that a Mithril snapshot Download has started.

mithril_extract_finished 🔒

Record when DL finished, if it fails, set size to None, otherwise the size of the downloaded file.

mithril_extract_started 🔒

Record that extracting the mithril snapshot archive has started.

mithril_sync_failure 🔒

Record when Mithril Cert validation starts, ends or fails).

mithril_validation_state 🔒

Record when Mithril Cert validation starts, ends or fails).

new_live_block 🔒

Count the validly deserialized blocks

new_mithril_update 🔒

Track the end of the current mithril update

peer_connected 🔒

Track statistics about connections to the cardano peer node.

rollback 🔒

Count a rollback

rollbacks 🔒

Extract the current rollback stats as a vec.

rollbacks_reset 🔒

Reset ALL the rollback stats for a given blockchain.

stats_invalid_block 🔒

Count the invalidly deserialized blocks

sync_started 🔒

Record when we started syncing

tip_reached 🔒

Record when we first reached tip. This can safely be called multiple times. Except for overhead, only the first call will actually record the time.

Type Aliases

RollbackMap 🔒

Record of rollbacks.

RollbackRecords 🔒

Individual rollback records.

RollbackTypeMap 🔒

Rollback Records per rollback type.

StatsMap 🔒

Type we use to manage the Sync Task handle map.