Struct cardano_chain_follower::ChainFollower

pub struct ChainFollower {
    chain: Network,
    end: Point,
    previous: Point,
    current: Point,
    fork: u64,
    snapshot: MithrilSnapshot,
    mithril_follower: Option<MithrilSnapshotIterator>,
    mithril_tip: Option<Point>,
    sync_updates: Receiver<Kind>,
}

The Chain Follower

Fields
chain: Network

The Blockchain network we are following.

end: Point

Where we end following.

previous: Point

Block we processed most recently.

current: Point

Where we are currently in the following process.

fork: u64

What fork were we last on

snapshot: MithrilSnapshot

Mithril Snapshot

mithril_follower: Option<MithrilSnapshotIterator>

Mithril Snapshot Follower

mithril_tip: Option<Point>

Mithril TIP Reached

sync_updates: Receiver<Kind>

Live Block Updates

Implementations

impl ChainFollower

pub async fn new(chain: Network, start: Point, end: Point) -> Self

Follow a blockchain.

Arguments

chain - The blockchain network to follow.

start - The point or tip to start following from (inclusive).

end - The point or tip to stop following from (inclusive).

Returns

The Chain Follower that will return blocks in the requested range.

Notes

IF end < start, the follower will immediately yield no blocks. IF end is TIP, then the follower will continue to follow even when TIP is reached. Otherwise only blocks in the request range will be returned.

Also, UNLIKE the blockchain itself, the only relevant information is the Slot#. The Block hash is not considered. If start is not an exact Slot#, then the NEXT Slot immediately following will be the first block returned. If the end is also not an exact Slot# with a block, then the last block will be the one immediately proceeding it.

To ONLY follow from TIP, set BOTH start and end to TIP.

async fn next_from_mithril(&mut self) -> Option<ChainUpdate>

If we can, get the next update from the mithril snapshot.

async fn next_from_live_chain(&mut self) -> Option<ChainUpdate>

If we can, get the next update from the mithril snapshot.

fn update_current(&mut self, update: &Option<ChainUpdate>) -> bool

Update the current Point, and return false if this fails.

pub(crate) async fn unprotected_next(&mut self) -> Option<ChainUpdate>

This is an unprotected version of next() which can ONLY be used within this crate. Its purpose is to allow the chain data to be inspected/validate prior to unlocking it for general access.

This function can NOT return None, but that state is used to help process data.

This function must not be exposed for general use.

pub async fn next(&mut self) -> Option<ChainUpdate>

Get the next block from the follower. Returns NONE is there is no block left to return.

pub async fn get_block(chain: Network, point: Point) -> Option<ChainUpdate>

Get a single block from the chain by its point.

If the Point does not point exactly at a block, it will return the next consecutive block.

This is a convenience function which just used ChainFollower to fetch a single block.

pub async fn get_tips(chain: Network) -> (Point, Point)

Get the current Immutable and live tips.

Note, this will block until the chain is synced, ready to be followed.

pub async fn post_txn(chain: Network, txn: TxBody) -> TxId

Schedule a transaction to be posted to the blockchain.

Arguments

chain - The blockchain to post the transaction on.

txn - The transaction to be posted.

Returns

TxId - The ID of the transaction that was queued.

pub async fn txn_sent(chain: Network, id: TxId) -> bool

Check if a transaction, known by its TxId, has been sent to the Peer Node.

Note, the TxId can ONLY be checked for ~6 hrs after it was posted. After which, it should be on the blockchain, and its the applications job to track if a transaction made it on-chain or not.