Struct cardano_chain_follower::Follower

pub struct Follower {
    client_connect_info: ClientConnectInfo,
    chain_update_rx: Receiver<Result<ChainUpdate>>,
    follow_task_request_tx: Sender<SetReadPointerRequest>,
    follow_task_join_handle: JoinHandle<()>,
    mithril_snapshot: Option<MithrilSnapshot>,
}

Cardano chain follower.

Fields
client_connect_info: ClientConnectInfo

Client connection information.

This is used to open more connections when needed.

chain_update_rx: Receiver<Result<ChainUpdate>>

Chain update receiver.

follow_task_request_tx: Sender<SetReadPointerRequest>

Follow task request sender.

follow_task_join_handle: JoinHandle<()>

Follow task thread join handle.

mithril_snapshot: Option<MithrilSnapshot>

Optional Mithril snapshot information.

Implementations

impl Follower

pub async fn connect( address: &str, network: Network, config: FollowerConfig, ) -> Result<Self>

Connects the follower to a producer using the node-to-node protocol.

Arguments

address: Address of the node to connect to.

network: The Network the client is assuming it’s connecting to.

config: Follower’s configuration (see FollowerConfigBuilder).

Errors

Returns Err if the connection could not be established.

pub async fn set_read_pointer<P>(&self, at: P) -> Result<Option<Point>>
where P: Into<PointOrTip>,

Set the follower’s chain read-pointer. Returns None if the point was not found on the chain.

Arguments

at: Point at which to set the read-pointer.

Errors

Returns Err if something went wrong while communicating with the producer.

pub fn read_block<P>(&self, at: P) -> ReadBlock
where P: Into<PointOrTip>,

Requests the client to read a block.

Arguments

at: Point at which to read the block.

pub fn read_block_range<P>(&self, from: Point, to: P) -> ReadBlockRange
where P: Into<PointOrTip>,

Request the client to read a block range.

Arguments

from: Block range start.

to: Block range end.

pub async fn next(&mut self) -> Result<ChainUpdate>

Receive the next chain update from the producer.

Errors

Returns Err if any producer communication errors occurred.

pub async fn close(self) -> Result<(), JoinError>

Closes the follower connection and stops its background task.

Errors

Returns Err if some error occurred in the background task.