Crate cardano_chain_follower

Cardano chain follower.

Re-exports

pub use metadata as Metadata;

Modules

chain_sync 🔒

Sync from the chain to an in-memory buffer.

chain_sync_config 🔒

Cardano chain sync configuration.

chain_sync_live_chains 🔒

Storage of each Live Chain per Blockchain.

chain_sync_ready 🔒

Flag to control if chain sync for a blockchain is ready. Can not consume the blockchain data until it is.

chain_update 🔒

An update of a blockchain

error 🔒

Library Crates Defined Errors

follow 🔒

Cardano chain follow module.

metadata

Metadata decoding and validating.

mithril_query 🔒

Functions which query or interact with the Immutable blockchain on disk.

mithril_snapshot 🔒

Internal Mithril snapshot functions.

mithril_snapshot_config 🔒

Configuration for the Mithril Snapshot used by the follower.

mithril_snapshot_data 🔒

Data about the current Mithril Snapshot

mithril_snapshot_iterator 🔒

Internal Mithril snapshot iterator functions.

mithril_snapshot_sync 🔒

Internal Mithril snapshot downloader task.

mithril_turbo_downloader 🔒

Turbo Downloads for Mithril Snapshots.

multi_era_block_data 🔒

Multi Era CBOR Encoded Block Data

network 🔒

Enum of possible Cardano networks.

point 🔒

A Cardano Point on the Blockchain.

snapshot_id 🔒

Simple ID for a mithril snapshot path known by its largest immutable file number

stats 🔒

Cardano Chain Follower Statistics

turbo_downloader

Serializable Parallel Download Processor

utils 🔒

Simple general purpose utility functions.

witness 🔒

Transaction Witness

Structs

ChainFollower

The Chain Follower

ChainSyncConfig

A Follower Connection to the Cardano Network.

ChainUpdate

Actual Chain Update itself.

MultiEraBlock

Multi-era block.

Point

A specific point in the blockchain. It can be used to identify a particular location within the blockchain, such as the tip (the most recent block) or any other block. It has special kinds of Point, available as constants: TIP_POINT, and ORIGIN_POINT.

Statistics

Statistics for a single follower network.

Enums

Kind

Enum of chain updates received by the follower.

Network

Enum of possible Cardano networks.

Constants

ORIGIN_POINT

The origin of the blockchain. It is used when the interest is in the very first point of the blockchain, regardless of its specific details.

TIP_POINT

The tip of the blockchain at the current moment. It is used when the specific point in the blockchain is not known, but the interest is in the most recent block (the tip). The tip is the point where new blocks are being added.

Type Aliases

Result

Crate result type.